python-llm/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
langchain==0.1.17
langchain-openai==0.1.4
langchain-community==0.0.36
pydantic==1.10.14
python-dotenv==1.0.1
